backend/app/requirements.txt
anyio==3.6.2
avro==1.11.1
avrotool==0.0.2
click==8.1.3
confluent-kafka==2.0.2
fastapi==0.91.0
fastavro==1.7.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
#mariadb==1.1.5.post3
packaging==23.0
paramiko~=3.1.0
peppercorn==0.6
pydantic==1.10.4
python-snappy==0.6.1
rec-avro==0.0.4
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
tinaa-logger~=3.0
tinaa-utils
typing_extensions==4.4.0
uvicorn==0.20.0
